app/requirements.txt
streamlit==1.37.1
langchain>=0.2.15
langchain-core>=0.2.10,<0.4.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.2.2
langchain-community
python-dotenv==0.21.0
requests==2.32.3
numpy==1.26.4
sentence-transformers==2.2.2
bs4